project folder structure updated with .py files
