+++ .env
@@ -1,0 +1,4 @@
+# API's keys
+
+SPOON_API_KEY = 5cf4d9752bea4c38b962e643ca227e27
+ROBO_API_KEY = qhSk7QdaM3p1YIzIdrPZ
