+++ .envrc
@@ -1,0 +1,2 @@
+
+dotenv
