+++ .gitignore
@@ -1,1 +1,2 @@
 raw_data/
+__pycache__
